access-control-service/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
kafka-python==2.0.2
prometheus-client==0.19.0
pydantic==2.5.0
SQLAlchemy==2.0.23
requests==2.31.0
